+++ rattler/settings/base.py
@@ -165,7 +165,7 @@ STATICFILES_DIRS = [
 
 # Authentication
 AUTH_USER_MODEL = 'user.User'
-#AUTHENTICATION_BACKENDS = ['path.to.class.YourBackend']
+AUTHENTICATION_BACKENDS = ['apps.register.NewModelBackend.NewModelBackend']
 
 LOGIN_URL = '/login/'
 LOGIN_REDIRECT_URL = '/'
